+++ packages/loki/libs/utils/index.d.ts
@@ -1,0 +1,6 @@
+export declare const cls: (preCls: any, className: any) => any;
+declare const _default: {
+    px2rem: ($px: any) => string;
+    cls: (preCls: any, className: any) => any;
+};
+export default _default;