+++ packages/loki/src/components/Gashapon/index.tsx
@@ -1,0 +1,22 @@
+// /**
+//  * @name 扭蛋机
+//  * @description
+//  *   可配置化扭蛋机
+//  *   1.扭蛋机背景
+//  *   2.机盒背景
+//  *   3.扭蛋球
+//  *   4.开始按钮 : 指针图片,大小
+//  *   5.落蛋口
+//  **/
+//  import React, { useState, useEffect, useRef } from 'react';
+//  import Utils from 'utils/index';
+
+//  const Gashapon = function() {
+//   return (
+//     <div className="loki-gashapon">
+//       gashapon
+//     </div>
+//   )
+//  }
+
+//  export default Gashapon;