+++ packages/ossa/src/components/carousel/index.tsx
@@ -67,7 +67,7 @@ function onChangeCarousel(
 ) {
   const { onChange } = props;
   setActiveIndex(e.detail.current);
-  onChange && onChange(e.detail.current);
+  onChange?.(e.detail.current);
 }
 
 function onClickCarousel(e: ITouchEvent, props: OsCarouselProps, currentIndex = 0) {
@@ -126,37 +126,38 @@ export default function Carousel(props: OsCarouselProps) {
         className='ossa-carousel-swiper'
       >
         {data.map((item: any, index: number) => (
-          <SwiperItem
-            className={`ossa-carousel-swiper-item ${
-              vertical
-                ? "ossa-carousel-swiper__item--vertical"
-                : "ossa-carousel-swiper__item"
-            }`}
-            style={
-              (gap &&
-                type === "image" && {
-                  padding: vertical
-                    ? `${Taro.pxTransform(gapWidth / 2)} 0`
-                    : `0 ${Taro.pxTransform(gapWidth / 2)}`,
-                }) ||
-              {}
-            }
-            key={item.content}
-          >
-            {type === "image" ? (
-              <Image
-                src={item.content}
-                className={`ossa-carousel-swiper__item__image ${item.className}`}
-                style={item.style}
-              ></Image>
-            ) : (
-              <View
-                className={`ossa-carousel-swiper__item__text ${item.className}`}
-                style={item.style}
-              >
-                {item.content}
-              </View>
-            )}
+          <SwiperItem key={item.content}>
+            <View
+              className={`ossa-carousel-swiper-item ${
+                vertical
+                  ? "ossa-carousel-swiper__item--vertical"
+                  : "ossa-carousel-swiper__item"
+              }`}
+              style={
+                (gap &&
+                  type === "image" && {
+                    padding: vertical
+                      ? `${Taro.pxTransform(gapWidth / 2)} 0`
+                      : `0 ${Taro.pxTransform(gapWidth / 2)}`,
+                  }) ||
+                {}
+              }
+            >
+              {type === "image" ? (
+                <Image
+                  src={item.content}
+                  className={`ossa-carousel-swiper__item__image ${item.className}`}
+                  style={item.style}
+                ></Image>
+              ) : (
+                <View
+                  className={`ossa-carousel-swiper__item__text ${item.className}`}
+                  style={item.style}
+                >
+                  {item.content}
+                </View>
+              )}
+            </View>
           </SwiperItem>
         ))}
       </Swiper>
@@ -188,7 +189,7 @@ Carousel.defaultProps = {
   interval: 4000,
   circular: true,
   indicatorDots: false,
-  indicatorColor: "#ffffff",
+  indicatorColor: "rgba(0, 0, 0, 0.3)",
   indicatorActiveColor: "#000000",
   current: 0,
   gap: false,