+++ packages/ossa-doc/CHANGELOG.md
@@ -1,5 +1,47 @@
 # ossa-doc
 
+## 0.1.0
+
+### Minor Changes
+
+- [#87](https://github.com/NeteaseYanxuan/OSSA/pull/87) [`3fc08d3`](https://github.com/NeteaseYanxuan/OSSA/commit/3fc08d3e75a6ce87ad9b4f19eaed67f3b10633ea) Thanks [@EchoHub](https://github.com/EchoHub)! - 营销组件 loki1.0 功能上线
+
+## 0.0.6
+
+### Patch Changes
+
+- [#78](https://github.com/NeteaseYanxuan/OSSA/pull/78) [`1039034`](https://github.com/NeteaseYanxuan/OSSA/commit/10390343772e0edacdf165f27866323b46dc2241) Thanks [@kiner-tang](https://github.com/kiner-tang)! - feat(noticeBar): 通知栏自定义滚动时长支持
+
+## 0.0.5
+
+### Patch Changes
+
+- [#73](https://github.com/NeteaseYanxuan/OSSA/pull/73) [`417eaf9`](https://github.com/NeteaseYanxuan/OSSA/commit/417eaf98883958eba7bbfa4d1ccc664c14b0a30e) Thanks [@kiner-tang](https://github.com/kiner-tang)! - docs: 删除列表组件示例中的冗余代码
+
+* [#79](https://github.com/NeteaseYanxuan/OSSA/pull/79) [`713303e`](https://github.com/NeteaseYanxuan/OSSA/commit/713303e1a73b07911f7fd45f4b7e9d40167fa5bd) Thanks [@kiner-tang](https://github.com/kiner-tang)! - feat: api 命名规范化 & 增加安全使用过期属性方法
+
+- [#67](https://github.com/NeteaseYanxuan/OSSA/pull/67) [`d9d5770`](https://github.com/NeteaseYanxuan/OSSA/commit/d9d57707f6439c98cf1f3d08c51b44b397cf0856) Thanks [@kiner-tang](https://github.com/kiner-tang)! - docs: 删除冗余输入框类型
+
+* [#75](https://github.com/NeteaseYanxuan/OSSA/pull/75) [`f7ac441`](https://github.com/NeteaseYanxuan/OSSA/commit/f7ac44193912383a94a2464fb08abf8543b89b35) Thanks [@kiner-tang](https://github.com/kiner-tang)! - docs: navBar 组件中 rightSlot 属性描述有误
+
+- [#72](https://github.com/NeteaseYanxuan/OSSA/pull/72) [`05093be`](https://github.com/NeteaseYanxuan/OSSA/commit/05093be98e4ad5171740303c81c147e95e23310c) Thanks [@kiner-tang](https://github.com/kiner-tang)! - feat: input 组件是否显示字符计数器属性名使用 showCount 属性替代
+
+* [#71](https://github.com/NeteaseYanxuan/OSSA/pull/71) [`b2a12cd`](https://github.com/NeteaseYanxuan/OSSA/commit/b2a12cded17a089fbb4e68b7a32ae8c1104b3931) Thanks [@kiner-tang](https://github.com/kiner-tang)! - docs: 补充 showSplitLine 属性文档说明
+
+- [#70](https://github.com/NeteaseYanxuan/OSSA/pull/70) [`c590e34`](https://github.com/NeteaseYanxuan/OSSA/commit/c590e34784ed2e9dacbfcec76a0f6dc9c53a651d) Thanks [@kiner-tang](https://github.com/kiner-tang)! - docs: 修改文档中 input 组件错误属性类型描述
+
+* [#74](https://github.com/NeteaseYanxuan/OSSA/pull/74) [`543602e`](https://github.com/NeteaseYanxuan/OSSA/commit/543602e11378405f6dda050e579197b7f96708a6) Thanks [@kiner-tang](https://github.com/kiner-tang)! - docs: 补充列表组件缺少的`leftIconSize`属性文档描述
+
+## 0.0.4
+
+### Patch Changes
+
+- [#61](https://github.com/NeteaseYanxuan/OSSA/pull/61) [`fd2910a`](https://github.com/NeteaseYanxuan/OSSA/commit/fd2910a9db05fb95d9f3b66b0c767c5e3ade3ebf) Thanks [@kiner-tang](https://github.com/kiner-tang)! - feat(Carousel): 走马灯点击回调支持传入当前点击子项索引
+
+* [#60](https://github.com/NeteaseYanxuan/OSSA/pull/60) [`752dd04`](https://github.com/NeteaseYanxuan/OSSA/commit/752dd043683877dac417e311cdbfd3bd1175154e) Thanks [@kiner-tang](https://github.com/kiner-tang)! - indicatorColor 与默认保持一致
+
+- [#58](https://github.com/NeteaseYanxuan/OSSA/pull/58) [`6d50ebe`](https://github.com/NeteaseYanxuan/OSSA/commit/6d50ebe652716ac2fb3a6f93107271b98d06034b) Thanks [@kiner-tang](https://github.com/kiner-tang)! - fix(Badge): 修复 Badge 的 info 为 0 或为空时显示异常问题
+
 ## 0.0.3
 
 ### Patch Changes